requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
motor==3.3.2
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.7
numpy==1.24.3
pydantic==2.4.2
python-multipart==0.0.6
httpx==0.25.1
tabulate==0.9.0